intelligentfleetmanagement/src/requirements.txt
#-r requirements-qampo.txt

absl-py==1.0.0
asttokens==2.0.5
backcall==0.2.0
bokeh==2.4.2
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dash[diskcache]==2.3.1
dash-bootstrap-components==1.1.0
dash-bootstrap-templates==1.0.5
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==0.8.3
Flask==2.1.1
Flask-Compress==1.11
idna==3.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.22.3
numpy-financial==1.0.0
openpyxl==3.0.9
ortools==9.3.10497
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
prompt-toolkit==3.0.29
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.3.24
stack-data==0.2.0
tenacity==8.0.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.0
dash-daq==0.5.0
--extra-index-url https://pypi.internal.2021.services/simple 
pgrace==0.4.5
datasources==1.8.2
scikit-learn==1.1.0
